SiteGadget-main/requirements.txt
urllib3==1.24.3
prettytable==0.7.2
anytree==2.8.0
requests==2.24.0
beautifulsoup4==4.9.3
dnspython==2.0.0
